Remove a reference to an invalid repo.
For one, this repo is actually https://github.com/StartTheShift/titan-tools but
was automatically updated to https://github.com/StartTheShift/janusgraph-tools
during the Great Renaming process, so it's already incorrect.

That said, it was last updated in 2013 and claims to only work with Titan 0.2.1,
so I think it's safe to assume it may not work with either recent versions of
Titan or any versions of JanusGraph, so it may be best to avoid pointing users
at it to avoid data corruption.

Signed-off-by: Misha Brukman <mbrukman@google.com>