Fixes #640: uncomment the Solr dependency.
The comment which hides Solr says that the dependency is not valid because it
requires Hadoop 2 while TinkerPop 3 only works with Hadoop 1.

JanusGraph currently depends on TinkerPop 3.2.6; per the docs:
http://tinkerpop.apache.org/docs/3.2.6/upgrade/
"Hadoop1 is no longer supported. Hadoop2 is now the only supported Hadoop
version in TinkerPop."

Thus, it should be safe to uncomment this dependency.

Signed-off-by: Misha Brukman <mbrukman@google.com>