Remove favicon.ico from docs. [skip ci]
The presence of the file in this repository means that it gets copied to every
new versioned docs directory as demonstrated in
https://github.com/JanusGraph/docs.janusgraph.org/pull/25

This file has been manually copied to the janusgraph.org and docs.janusgraph.org
repos at the top-level, so they will show up for browsers automatically there.

We don't need to keep a copy of this file here, as it does not serve any value.

Signed-off-by: Misha Brukman <mbrukman@google.com>